Developers Window implemented.
