+++ Tutti/Applications/TouchUI/Views/DevelopersWindow.xaml.cs
@@ -1,0 +1,27 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Shapes;
+
+namespace TouchUI.Views
+{
+    /// <summary>
+    /// Interaction logic for DevelopersWindow.xaml
+    /// </summary>
+    public partial class DevelopersWindow : Window
+    {
+        public DevelopersWindow()
+        {
+            InitializeComponent();
+        }
+    }
+}
